toblse-tlv: fix segfault from OP_GET printing
Segmentation faults occured when trying to OP_GET from a legacy EEPROM.
The current code tracked `key_len` as a flag, if it was 0, then it
pointed to static string from `tlv_code_list`, if it was not 0, then it
indicated allocation. Previously, the code attempted to `realloc` to a
read-only string, causing segmentation fault.

ref: https://lab.8devices.com/firmux/tlvstore/-/issues/2